:sparkes: Creates the factorial route

+++ src/routes.ts
@@ -8,7 +8,10 @@ const route = Router();
 const factorialController = new FactorialController();
 
 route.get('/factorial', [
-    query('number').isNumeric().withMessage('Você precisa fornecer um número')
+    query('number')
+        .isNumeric()
+        .withMessage('Você precisa fornecer um número')
+        .toInt()
 ], factorialController.compute);
 
 export default route;
